added some better code architecture for utils and accessor methods
